Create and implement custom hook to handle wine storing
