+++ client/src/components/App/App.js
@@ -5,32 +5,10 @@ import Header from 'components/Header/Header'
 import WineListing from 'components/WineListing/WineListing'
 import WineStorage from 'components/WineStorage/WineStorage'
 import { Route, Switch } from 'react-router-dom'
-import useLocalStorage from 'hooks/useLocalStorage'
+import useHandleSave from 'hooks/useHandleSave'
 
 export default function App() {
-  const [savedWines, setSavedWines] = useLocalStorage('wines', [])
-
-  function saveWine(target) {
-    const newSavedWine = target
-    setSavedWines([newSavedWine, ...savedWines])
-  }
-  function removeWine(target) {
-    const updatedSavedWines = savedWines.filter(wine => wine.id !== target.id)
-    setSavedWines([...updatedSavedWines])
-  }
-
-  function toggleSave(target) {
-    const includesTarget =
-      savedWines !== undefined
-        ? savedWines.some(item => item.id === target.id)
-        : false
-    console.log(includesTarget)
-    if (includesTarget) {
-      removeWine(target)
-    } else {
-      saveWine(target)
-    }
-  }
+  const [savedWines, toggleSave] = useHandleSave()
 
   return (
     <Grid>
@@ -38,11 +16,7 @@ export default function App() {
       <Main>
         <Switch>
           <Route path="/wine-storage">
-            <WineStorage
-              savedWines={savedWines}
-              onSaveToggle={toggleSave}
-              setSavedWines={setSavedWines}
-            />
+            <WineStorage savedWines={savedWines} onSaveToggle={toggleSave} />
           </Route>
           <Route path="/">
             <WineListing
